+++ packages/algob/src/index.ts
@@ -1,5 +1,6 @@
 import { addressToPk, stringToBytes, uint64ToBigEndian } from "@algo-builder/runtime";
 
+import * as ERRORS from "./errors/errors-list";
 import {
   createMsigAddress,
   loadAccountsFromEnv,
@@ -18,6 +19,7 @@ import * as runtime from "./runtime";
 import * as types from "./types";
 const convert = { uint64ToBigEndian, addressToPk, stringToBytes };
 export {
+  ERRORS,
   mkAccounts,
   createMsigAddress,
   loadAccountsFromFile,