+++ packages/algob/test/builtin-tasks/run.ts
@@ -4,7 +4,7 @@ import fs from "fs";
 import { splitAfter } from "../../src/builtin-tasks/run";
 // import * as fsExtra from "fs-extra";
 import { TASK_DEPLOY, TASK_RUN } from "../../src/builtin-tasks/task-names";
-import { ERRORS } from "../../src/internal/core/errors-list";
+import { ERRORS } from "../../src/errors/errors-list";
 import { loadCheckpoint } from "../../src/lib/script-checkpoints";
 import { useEnvironment } from "../helpers/environment";
 import { expectBuilderErrorAsync } from "../helpers/errors";