+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 + Added support for sub directories in assets folder, with tests.
 + TEALv3 support in `@algo-builder/runtime`: `assert`, `pushint`, `pushbytes`, `swap` opcodes.
 + Update runtime to process execParams.deployASA, deploySSC, OptInToASA, OptIntoSSC
++ Exported `@algorand-builder/algob`, `@algorand-builder/runtime` error types and make it accessible for API documentation.
 
 ### Bug fixes